SERVER: Receving SIGSEGV process on shutdown
There is race condition when dynamic libraries are unloaded. Talloc
library calls our destructors but they still need openssl calls
which might be not available.

Solution is to free explicitely memory context and trigger
destructors before calling exit(). In this PR the SIGTERM
handler is moved from individual providers to generel
backend code.

Also generic server code is changed to explicitely free
memory context when signal is received.

Resolves:
https://bugzilla.redhat.com/show_bug.cgi?id=1672584
